fix: sonnar ouath 3
